Add multi-select for tag only editing mode
This adds multi-select support for tag only editing mode

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/2472

Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/522
